Added furniture manager and furniture script
made the furniture manager to track the amount of furniture on the level, all the ruined furniture and also made a furniture script that should update the furniture manager if collided with the player, although not been able to test that. Also set the ability for the furniture to be reset if needed.