Added Camershake and fixed enemy walkling animation
Added camera shake which allows, through code, someone to add shake with a duration and magnitude. Also fixed the animation for enemies facing the wrong way, wasn't sure why the issue occurred but since it was always facing the wrong way I just took the opposite of the current Boolean used to set it and took the opposite and that fixed it.